server/app/requirements.txt
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
Authlib==1.2.0
cffi==1.15.1
click==8.1.3
cryptography==38.0.4
dnspython==2.2.1
Flask==2.2.2
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
pycparser==2.21
pymongo==4.3.3
Werkzeug==2.2.2
